feat(agent): reorganize backend into module with separate files
- Move backend code to backends/ directory structure
- Split ClaudeBackend into backends/claude.rs
- Split CursorBackend into backends/cursor.rs
- Add backends/mod.rs with Backend trait and registry
- Simplify spawn.rs to use new backend module

Closes #127
